move selectors to its own dir
